AI Moving
- AI move to adjust arrow get closer to player.
- AI Move to get away from player's arrow

+++ include/Maths.h
@@ -24,8 +24,8 @@ public:
 class Point {
 public:
   Point(double x = 0, double y = 0);
-  void setX(int x);
-  void setY(int y);
+  void setX(double x);
+  void setY(double y);
 
   double x, y;
 };
